smartbugs/tests/get-source-code/pip-requirements.txt
py-etherscan-api
requests>=2.20.0
typing==3.6.4
